Create admin-page-init.php
